refactor: moves logic for methods to a single model
