+++ front/package.json
@@ -18,7 +18,7 @@
     "extends": "../package.json"
   },
   "dependencies": {
-    "@apollo/client": "^3.5.10",
+    "@apollo/client": "^3.6.5",
     "@fluentui/react-icons": "^2.0.170-rc.7",
     "@griffel/react": "^1.0.5",
     "@headlessui/react": "^1.6.2",
@@ -35,7 +35,7 @@
     "@types/mime": "^2.0.3",
     "@types/react-datepicker": "^4.4.1",
     "apollo-upload-client": "^17.0.0",
-    "base-analysis": "^0.1.3",
+    "base-analysis": "^0.1.5",
     "byte-size": "^8.1.0",
     "clsx": "^1.1.1",
     "date-fns": "^2.28.0",
@@ -55,12 +55,12 @@
     "react-dom": "^17.0.2",
     "react-dropzone": "^14.2.1",
     "react-error-boundary": "^3.1.4",
-    "react-hook-form": "^7.31.1",
+    "react-hook-form": "^7.31.2",
     "react-kbs": "^2.0.0",
-    "react-plot": "^0.16.0",
+    "react-plot": "^0.19.1",
     "react-popper": "^2.3.0",
-    "react-query": "^3.34.20",
-    "react-router-dom": "^6.2.2",
+    "react-query": "^3.39.0",
+    "react-router-dom": "^6.3.0",
     "react-wafer": "^2.3.0",
     "slate": "^0.72.8",
     "slate-history": "^0.66.0",
@@ -89,17 +89,17 @@
     "@zakodium/tailwind-config": "^3.1.2",
     "autoprefixer": "^10.4.7",
     "concurrently": "^7.0.0",
-    "eslint": "^8.6.0",
-    "eslint-import-resolver-typescript": "^2.5.0",
-    "eslint-plugin-import": "^2.25.4",
+    "eslint": "^8.16.0",
+    "eslint-import-resolver-typescript": "^2.7.1",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-react": "^7.30.0",
+    "eslint-plugin-react-hooks": "^4.5.0",
     "jscpd": "^3.4.5",
     "postcss": "^8.4.14",
-    "prettier": "2.5.1",
+    "prettier": "2.6.2",
     "react-test-renderer": "^17.0.2",
-    "typescript": "^4.5.5",
+    "typescript": "^4.7.2",
     "vite": "^2.8.6",
     "vite-tsconfig-paths": "^3.4.1"
   }
