+++ front/src/index.tsx
@@ -11,9 +11,9 @@ import List from '@/addons/fileSync/pages/list';
 import NoMatch from '@/pages/404';
 import App from '@/pages/App';
 import Login from '@/pages/login';
-import MeasurementCreate from '@/pages/measurements/create/Default';
-import MeasurementDetail from '@/pages/measurements/transfer/detail';
-import MeasurementsList from '@/pages/measurements/transfer/list';
+import MeasurementCreate from '@/pages/measurements/Create';
+import MeasurementDetail from '@/pages/measurements/Detail';
+import MeasurementsList from '@/pages/measurements/List';
 import NotebookCreate from '@/pages/notebook/Create';
 import NotebookDetail from '@/pages/notebook/Detail';
 import NotebookList from '@/pages/notebook/list';
