+++ front/src/models/measurements.ts
@@ -1,0 +1,6 @@
+import { MeasurementTypes } from '@/generated/graphql';
+
+export const measurementTypeList = [
+  MeasurementTypes.TRANSFER,
+  MeasurementTypes.XRAY,
+];
