+++ front/src/models/sample.ts
@@ -1,0 +1,2 @@
+export const sampleLevels = ['wafer', 'sample', 'dye', 'device'];
+export const sampleLevelsStrict = ['wafer', 'sample', 'dye', 'device'] as const;
