+++ front/src/pages/measurements/List/MeasurementPlot/MeasurementTypeRender.tsx
