+++ front/src/pages/samples/list/Device.tsx
@@ -19,6 +19,7 @@ import {
   Variant,
 } from '@/components/tailwind-ui';
 import { Sample } from '@/generated/graphql';
+import { sampleLevels } from '@/models/sample';
 
 import SamplesList from './Default';
 
@@ -30,7 +31,7 @@ export default function DeviceList() {
       <div className="col-span-3">
         <SamplesList
           kind="device"
-          levels={['wafer', 'sample', 'dye', 'device']}
+          levels={sampleLevels}
           action={
             <TableQuery.ActionsColumn>
               {(row) => (
