+++ front/src/pages/samples/list/Dye.tsx
@@ -19,6 +19,7 @@ import {
   Variant,
 } from '@/components/tailwind-ui';
 import { Sample } from '@/generated/graphql';
+import { sampleLevels } from '@/models/sample';
 
 import SamplesList from './Default';
 
@@ -30,7 +31,7 @@ export default function DyeList() {
       <div className="col-span-3">
         <SamplesList
           kind="dye"
-          levels={['wafer', 'sample', 'dye']}
+          levels={sampleLevels.slice(0, 3)}
           action={
             <TableQuery.ActionsColumn>
               {(row) => (
