+++ front/src/pages/samples/list/Sample.tsx
@@ -21,6 +21,7 @@ import {
   Variant,
 } from '@/components/tailwind-ui';
 import { Sample } from '@/generated/graphql';
+import { sampleLevels } from '@/models/sample';
 
 import SamplesList from './Default';
 
@@ -32,7 +33,7 @@ export default function SampleList() {
       <div className="col-span-3">
         <SamplesList
           kind="sample"
-          levels={['wafer', 'sample']}
+          levels={sampleLevels.slice(0, 2)}
           action={
             <TableQuery.ActionsColumn>
               {(row) => (
